requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backoff==1.11.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
coverage==5.5
deprecation==2.1.0
eyed3==0.9.6
filetype==1.0.7
idna==3.2
multidict==5.1.0
packaging==21.0
pip==21.1.3
pydub==0.25.1
pyparsing==2.4.7
python-dotenv==0.18.0
requests==2.26.0
six==1.16.0
spotify==0.10.2
spotipy==2.18.0
toml==0.10.2
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wheel==0.36.2
wincertstore==0.2
yarl==1.6.3
